# backend/docs/specs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Beza Platform API'
  description: "مَنصة بَiza — API موحّد لفرق Flutter و USSD\nالمنصة المالية السورية رقم 1 للخدمات المالية الرقمية"
  version: 1.0.0
  contact:
    name: 'Beza Platform'
    email: api@beza-platform.com
servers:
  - url: 'http://localhost:8000/api'
    description: 'Local Development'
paths:
...
  'v1/agents/{id}':
    get:
      tags:
        - Agent
      summary: 'Get agent details'
      description: 'عرض بيانات الوكيل'
      operationId: getAgent
      parameters:
        -
          in: path
          name: id
          required: true
          schema:
            type: string
          description: 'Agent ULID'
      responses:
        200:
          description: 'Agent details'
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/Agent' } } }
        404:
          $ref: '#/components/responses/NotFoundError'
...
  v1/financial/transfer:
    post:
      tags:
        - 'Financial Core'
      summary: 'Transfer between wallets'
      description: 'تحويل أموال بين محفظتين — يستخدم X-Idempotency-Key لمنع التكرار'
      operationId: transfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        201:
          description: 'Transfer initiated'
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/Transaction' } } }
        401:
          $ref: '#/components/responses/UnauthorizedError'
        422:
          $ref: '#/components/responses/UnprocessableEntity'
  v1/financial/deposit:
    post:
      tags:
        - 'Financial Core'
      summary: 'Deposit to wallet'
      description: 'إيداع في محفظة'
      operationId: deposit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositRequest'
      responses:
        201:
          description: 'Deposit created'
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/Transaction' } } }
        401:
          $ref: '#/components/responses/UnauthorizedError'
        422:
          $ref: '#/components/responses/UnprocessableEntity'
  v1/financial/withdraw:
    post:
      tags:
        - 'Financial Core'
      summary: 'Withdraw from wallet'
      description: 'سحب من محفظة'
      operationId: withdraw
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawRequest'
      responses:
        201:
          description: 'Withdrawal created'
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/Transaction' } } }
        401:
          $ref: '#/components/responses/UnauthorizedError'
        422:
          $ref: '#/components/responses/UnprocessableEntity'
...
  'v1/financial/transactions/{id}':
    get:
      tags:
        - 'Financial Core'
      summary: 'Get transaction details'
      description: 'عرض تفاصيل معاملة'
      operationId: getTransaction
      parameters:
        -
          in: path
          name: id
          required: true
          schema:
            type: string
          description: 'Transaction ULID'
      responses:
        200:
          description: 'Transaction details'
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/Transaction' } } }
        404:
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT token issued at authentication'
    IdempotencyKey:
      type: apiKey
      in: header
      name: X-Idempotency-Key
      description: 'Idempotency key for safe retries (ULID format recommended)'
  schemas:
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        code:
          type: integer
          nullable: true
    ValidationErrorResponse:
      type: object
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    Agent:
      type: object
      properties:
        id:
          type: string
          description: ULID
        user_id:
          type: string
        phone:
          type: string
        name:
          type: string
        name_ar:
          type: string
        kyc_tier:
          type: string
          enum:
            - t0
            - t1
            - t2
            - t3
        status:
          type: string
          enum:
            - pending
            - active
            - suspended
            - terminated
        id_type:
          type: string
          nullable: true
        id_number:
          type: string
          nullable: true
        is_verified:
          type: boolean
        verified_at:
          type: string
          format: date-time
          nullable: true
        gps_lat:
          type: number
          format: float
          nullable: true
        gps_lng:
          type: number
          format: float
          nullable: true
        address:
          type: string
          nullable: true
        address_ar:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
...
    Transaction:
      type: object
      properties:
        id:
          type: string
          description: ULID
        type:
          type: string
          enum:
            - transfer
            - deposit
            - withdraw
            - reversal
            - settlement
            - fee
        status:
          type: string
          enum:
            - initiated
            - held
            - posted
            - settled
            - reversed
            - failed
        wallet_id:
          type: string
          nullable: true
        from_account_id:
          type: string
          nullable: true
        to_account_id:
          type: string
          nullable: true
        amount:
          type: integer
        currency:
          type: string
          enum:
            - SYP
            - USD
        fee_amount:
          type: integer
        fee_account_id:
          type: string
          nullable: true
        fee_basis_points:
          type: integer
        idempotency_key:
          type: string
        description:
          type: string
          nullable: true
        description_ar:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
        reversal_of:
          type: string
          nullable: true
        journal_entry_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TransferRequest:
      type: object
      required:
        - from_wallet_id
        - to_wallet_id
        - amount
        - currency
        - idempotency_key
      properties:
        from_wallet_id:
          type: string
          description: 'Source wallet ULID'
        to_wallet_id:
          type: string
          description: 'Destination wallet ULID'
        amount:
          type: integer
          minimum: 1
          description: 'Amount in minor units'
        currency:
          type: string
          enum:
            - SYP
            - USD
        idempotency_key:
          type: string
        fee:
          type: object
          properties:
            rule_id:
              type: string
              description: 'Fee rule ULID (required_without)'
            description:
              type: string
              nullable: true
            description_ar:
              type: string
              nullable: true
    DepositRequest:
      type: object
      required:
        - wallet_id
        - amount
        - currency
        - idempotency_key
      properties:
        wallet_id:
          type: string
          description: 'Wallet ULID'
        amount:
          type: integer
          minimum: 1
          description: 'Amount in minor units'
        currency:
          type: string
          enum:
            - SYP
            - USD
        idempotency_key:
          type: string
    WithdrawRequest:
      type: object
      required:
        - wallet_id
        - amount
        - currency
        - idempotency_key
      properties:
        wallet_id:
          type: string
          description: 'Wallet ULID'
        amount:
          type: integer
          minimum: 1
          description: 'Amount in minor units'
        currency:
          type: string
          enum:
            - SYP
            - USD
        idempotency_key:
          type: string
...
  responses:
    UnauthorizedError:
      description: 'Missing or invalid authentication'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: Unauthenticated
            code: 401
    UnprocessableEntity:
      description: 'Validation error'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            message: 'Validation failed'
            errors:
              amount: ['The amount field is required.']
    NotFoundError:
      description: 'Resource not found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: 'Resource not found'
            code: 404
security:
  -
    bearerAuth: []
...